[Extension] Rename dpsl.ts to telemetry-extension-types.ts
Rename the file to better reflect the content.

The telemetry-extension-types.ts file contains various type
definitions defined by the telemetry extension APIs.

BUG=b/314066623
