+++ common/telemetry-extension-types.ts
