+++ diagnostics-extension/src/services/events.ts
@@ -10,7 +10,10 @@
 // Allow type-casting into any for access.
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
-import {EventCategory, EventSupportStatusInfo} from '@common/dpsl';
+import {
+  EventCategory,
+  EventSupportStatusInfo,
+} from '@common/telemetry-extension-types';
 import {Response} from '@common/message';
 import * as fake_events from './fake_events.data';
 import {environment} from '../environments/environment';
